plugin-redfish/rfphandler/storage.go
//(C) Copyright [2020] Hewlett Packard Enterprise Development LP
//
//Licensed under the Apache License, Version 2.0 (the "License"); you may
//not use this file except in compliance with the License. You may obtain
//a copy of the License at
//
//    http://www.apache.org/licenses/LICENSE-2.0
//
//Unless required by applicable law or agreed to in writing, software
//distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
//WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
//License for the specific language governing permissions and limitations
// under the License.

//Package rfphandler ...
package rfphandler

import (
	"net/http"

	iris "github.com/kataras/iris/v12"
)

//CreateVolume function is used for creating a volume under storage
func CreateVolume(ctx iris.Context) {
	// This function is yet to be implemented
	ctx.StatusCode(http.StatusNotImplemented)
}
